deleted unused files (for real), changed to recent API, renamed folder
