+++ portal/autoconfig.py
@@ -104,7 +104,7 @@ SETTINGS = {
         "django.middleware.csrf.CsrfViewMiddleware",
         "django.contrib.auth.middleware.AuthenticationMiddleware",
         "django.contrib.auth.middleware.SessionAuthenticationMiddleware",
-        "portal.middleware.django_online_status.middleware.OnlineStatusMiddleware",
+        "portal.middleware.online_status.OnlineStatusMiddleware",
         "django.contrib.messages.middleware.MessageMiddleware",
         "django.middleware.clickjacking.XFrameOptionsMiddleware",
         "deploy.middleware.exceptionlogging.ExceptionLoggingMiddleware",
@@ -141,7 +141,7 @@ SETTINGS = {
 RELATIONSHIPS = [
     OrderingRelationship(
         "MIDDLEWARE",
-        "portal.middleware.django_online_status.middleware.OnlineStatusMiddleware",
+        "portal.middleware.online_status.OnlineStatusMiddleware",
         after=["django.contrib.auth.middleware.AuthenticationMiddleware"],
         add_missing=False,
     ),
