+++ portal/middleware/online_status/middleware.py
@@ -1,19 +1,23 @@
 from django.core.cache import cache
-from portal.middleware.django_online_status.status import (
-    refresh_user,
-    refresh_users_list,
-)
+from portal.middleware.online_status.status import refresh_user, refresh_users_list
 from status import (
     OnlineStatus,
     CACHE_PREFIX_USER,
     CACHE_PREFIX_ANONYM_USER,
     ONLY_LOGGED_USERS,
 )
-from django.utils.deprecation import MiddlewareMixin
 
 
-class OnlineStatusMiddleware(MiddlewareMixin):
+class OnlineStatusMiddleware(object):
     """Cache OnlineStatus instance for an authenticated User"""
+
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        self.process_request(request)
+        response = self.get_response(request)
+        return response
 
     def process_request(self, request):
         if request.user.is_authenticated():
